Commit 11
Figured out how we're going to read each line by using stringstream so that we can skip through specified characters and series of characters to obtain the data we want which in this case is the word, pos, and definition. We need a to use a struct and the vector data type to be a struct so that we can pass the neccessary data along. We will then use that data in the Dictionary.cpp to organize that data into each HashMap(one for each POS for easier searching).